Update blacklist to denylist
