--- blacklist.txt
@@ -1,31 +1,0 @@
-password
-admin
-administrator
-root
-January
-February
-March
-April
-May
-June
-July
-August
-September
-October
-November
-December
-spring
-summer
-fall
-autumn
-winter
-qwerty
-qwertyuiop
-1qazxsw23edc
-1qazxsw2
-xsw2zaq1
-xsw21qaz
-2023
-2024
-2025
-2026
